feat: 扩展 value_table bit 源

+++ include/protoscope/scripting/script_host.hpp
@@ -65,7 +65,7 @@ struct ValueTableBitValueLabels {
 struct ValueTableRowDescriptor {
     std::uint32_t id{0};
     std::uint32_t sourceId{0};
-    std::optional<std::uint8_t> bit;
+    std::optional<std::uint32_t> bit; // bit 展开行的位下标；整数源按 64 位读取，字节源按字节低位优先读取。
     std::string label;
     std::string unit;
     std::string note;
